Whole refinements.Still in midway.

+++ class/AAtext.h
@@ -7,8 +7,9 @@ namespace scgb{
     std::vector<wchar_t> data;
   public:
     void Draw();
-    AAtext(int x,int y,SizeData*,std::string);
+    AAtext(double x,double y,SizeData*,std::string);
   };
+  
   class AAtextBuilder:public BWBuilder{
   protected:
     std::string filename;
@@ -16,7 +17,7 @@ namespace scgb{
     Drawable* GetResult(){
       return new AAtext(x,y,psd,filename);
     }
-    AAtextBuilder(int x,int y,std::string fn)
+    AAtextBuilder(double x,double y,std::string fn)
       :BWBuilder(x,y,0,0){filename=fn;};
     virtual ~AAtextBuilder(){};
   };
